data/requirements.txt
# Requirements for parser
cdsapi < 0.8
numpy < 2.4
# Required by xarray to process NetCDF files
netcdf4 < 1.8
xarray < 2025.11

# Requirements for producer
kafka-python < 2.3
aws-msk-iam-sasl-signer-python < 1.1
python-dotenv < 1.2
